requirements.txt
dj-database-url==0.5.0
Django==2.2
gunicorn==20.1.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
unicorn==1.0.2
Unipath==1.1
whitenoise==5.2.0
